feat: improve code from code review
